Restructured library folder.
