+++ CMakeIncludes/CMakeMacros.txt
@@ -27,11 +27,12 @@ MACRO(ARDUINOPROJECT PROJECT_NAME)
 		COMMAND rm -rf ${CMAKE_CURRENT_SOURCE_DIR}/src
 		COMMAND mkdir ${CMAKE_CURRENT_SOURCE_DIR}/src
 		COMMAND mkdir ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib
-		COMMAND ln -s ${CMAKE_SOURCE_DIR}/arduino/*.cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib
-		COMMAND ln -s ${CMAKE_SOURCE_DIR}/arduino/*.h ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib
+		COMMAND mkdir ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib/arduino
+		COMMAND ln -s ${CMAKE_SOURCE_DIR}/lib/arduino/*.cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib/arduino
+		COMMAND ln -s ${CMAKE_SOURCE_DIR}/lib/arduino/*.h ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib/arduino
 		COMMAND ln -s ${CMAKE_SOURCE_DIR}/lib/*.cpp ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib
 		COMMAND ln -s ${CMAKE_SOURCE_DIR}/lib/*.h ${CMAKE_CURRENT_SOURCE_DIR}/src/iplib
-		COMMAND arduino-cli compile -b ${ARDUINO_FQBN} --output ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME} --build-path ${CMAKE_CURRENT_BINARY_DIR}/build ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_NAME}.ino
+		COMMAND arduino-cli compile -b ${ARDUINO_FQBN} --output ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME} --build-path ${CMAKE_CURRENT_BINARY_DIR}/build ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_NAME}.cpp
 	)
 ENDMACRO()
