+++ CMakeLists.txt
@@ -11,7 +11,7 @@ set(INVPEND_COMMON_LIBS iplib)
 
 message(STATUS "Setting up ${INVPEND_INCLUDE_ONLY_DIR} ...")
 file(MAKE_DIRECTORY ${INVPEND_INCLUDE_ONLY_DIR}/iplib)
-file(GLOB INVPEND_HEADER_FILES ${InvPend_SOURCE_DIR}/lib/*.h)
+file(GLOB INVPEND_HEADER_FILES ${InvPend_SOURCE_DIR}/lib/*.h ${InvPend_SOURCE_DIR}/lib/desktop/*.h)
 file(COPY ${INVPEND_HEADER_FILES} DESTINATION ${INVPEND_INCLUDE_ONLY_DIR}/iplib)
 message(STATUS "Done setting up ${INVPEND_INCLUDE_ONLY_DIR} ...")
