+++ lib/arduino/QuadEncoder.h
