+++ lib/arduino/QuadEncoderSensor.h
