+++ lib/desktop/AbstractSocket.h
