+++ lib/desktop/Address.cpp
@@ -1,5 +1,3 @@
-#ifndef ARDUINO
-
 #include <iostream>
 
 #include "Address.h"
@@ -34,5 +32,3 @@ ostream &operator<<(ostream &out, const Address &rhs) {
 
 } // net
 } // iplib
-
-#endif // ARDUINO
